requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alabaster==1.0.0
asgiref==3.9.1
attrs==25.3.0
babel==2.17.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
docutils==0.21.2
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
libsql-client==0.3.1
libsql-experimental==0.0.55
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
propcache==0.3.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.4
roman-numerals-py==3.1.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sqlalchemy-libsql==0.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
